+++ packages/core/src/RenderingEngine/VideoViewport.ts
@@ -172,9 +172,10 @@ class VideoViewport extends Viewport {
 
     let rowCosines = imagePlaneModule.rowCosines as Point3;
     let columnCosines = imagePlaneModule.columnCosines as Point3;
+    const usingDefaultValues = imagePlaneModule.usingDefaultValues;
 
     // if null or undefined
-    if (rowCosines == null || columnCosines == null) {
+    if (usingDefaultValues || rowCosines == null || columnCosines == null) {
       rowCosines = [1, 0, 0] as Point3;
       columnCosines = [0, 1, 0] as Point3;
     }
@@ -836,6 +837,9 @@ class VideoViewport extends Viewport {
       return false;
     }
     const match = referencedImageId.match(VideoViewport.frameRangeExtractor);
+    if (!match) {
+      return true;
+    }
     if (!match[2]) {
       return true;
     }
@@ -1130,16 +1134,17 @@ class VideoViewport extends Viewport {
   public addImages(stackInputs: IStackInput[]) {
     const actors = this.getActors();
     stackInputs.forEach((stackInput) => {
-      const image = cache.getImage(stackInput.imageId);
+      const { imageId, ...rest } = stackInput;
+      const image = cache.getImage(imageId);
 
       const imageActor = this.createActorMapper(image);
       const uid = stackInput.actorUID ?? uuidv4();
       if (imageActor) {
-        actors.push({ uid, actor: imageActor });
+        actors.push({ uid, actor: imageActor, referencedId: imageId, ...rest });
         if (stackInput.callback) {
           stackInput.callback({
             imageActor: imageActor as unknown as ImageActor,
-            imageId: stackInput.imageId,
+            imageId,
           });
         }
       }