+++ packages/core/src/RenderingEngine/helpers/createVolumeMapper.ts
@@ -2,7 +2,9 @@ import { vtkSharedVolumeMapper } from '../vtkClasses';
 import { getConfiguration } from '../../init';
 import type vtkImageData from '@kitware/vtk.js/Common/DataModel/ImageData';
 import type vtkOpenGLTexture from '@kitware/vtk.js/Rendering/OpenGL/Texture';
-import type vtkVolumeMapper from '@kitware/vtk.js/Rendering/Core/VolumeMapper';
+import vtkVolumeMapper from '@kitware/vtk.js/Rendering/Core/VolumeMapper';
+import vtkDataArray from '@kitware/vtk.js/Common/Core/DataArray';
+
 /**
  * Given an imageData and a vtkOpenGLTexture, it creates a "shared" vtk volume mapper
  * from which various volume actors can be created.
@@ -36,5 +38,34 @@ export default function createVolumeMapper(
   volumeMapper.setSampleDistance(sampleDistance);
   volumeMapper.setScalarTexture(vtkOpenGLTexture);
 
+  return volumeMapper;
+}
+
+/**
+ * Converts a shared mapper to a non-shared mapper. Sometimes we need to detach
+ * a shared mapper and apply some changes to it, since otherwise, the changes
+ * will be applied to all the mappers that share the same data.
+ *
+ * @param sharedMapper - The shared mapper to convert.
+ * @returns The converted volume mapper.
+ */
+export function convertMapperToNotSharedMapper(sharedMapper: vtkVolumeMapper) {
+  const volumeMapper = vtkVolumeMapper.newInstance();
+  volumeMapper.setBlendMode(sharedMapper.getBlendMode());
+
+  const imageData = sharedMapper.getInputData();
+  const { voxelManager } = imageData.get('voxelManager');
+  const values = voxelManager.getCompleteScalarDataArray();
+
+  const scalarArray = vtkDataArray.newInstance({
+    name: `Pixels`,
+    values,
+  });
+
+  imageData.getPointData().setScalars(scalarArray);
+
+  volumeMapper.setInputData(imageData);
+  volumeMapper.setMaximumSamplesPerRay(sharedMapper.getMaximumSamplesPerRay());
+  volumeMapper.setSampleDistance(sharedMapper.getSampleDistance());
   return volumeMapper;
 }