+++ packages/tools/src/utilities/math/index.ts
@@ -1,5 +1,6 @@
 import * as aabb from './aabb';
 import * as BasicStatsCalculator from './basic';
+import * as circle from './circle';
 import * as ellipse from './ellipse';
 import * as lineSegment from './line';
 import * as point from './point';
@@ -10,6 +11,7 @@ import * as vec2 from './vec2';
 export {
   aabb,
   BasicStatsCalculator,
+  circle,
   ellipse,
   lineSegment,
   point,