+++ packages/adapters/CHANGELOG.md
@@ -3,6 +3,84 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.4](https://github.com/dcmjs-org/dcmjs/compare/v2.14.3...v2.14.4) (2025-01-06)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.14.3](https://github.com/dcmjs-org/dcmjs/compare/v2.14.2...v2.14.3) (2025-01-06)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.14.2](https://github.com/dcmjs-org/dcmjs/compare/v2.14.1...v2.14.2) (2024-12-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.14.1](https://github.com/dcmjs-org/dcmjs/compare/v2.14.0...v2.14.1) (2024-12-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [2.14.0](https://github.com/dcmjs-org/dcmjs/compare/v2.13.0...v2.14.0) (2024-12-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [2.13.0](https://github.com/dcmjs-org/dcmjs/compare/v2.12.3...v2.13.0) (2024-12-20)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.12.3](https://github.com/dcmjs-org/dcmjs/compare/v2.12.2...v2.12.3) (2024-12-18)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.12.2](https://github.com/dcmjs-org/dcmjs/compare/v2.12.1...v2.12.2) (2024-12-18)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.12.1](https://github.com/dcmjs-org/dcmjs/compare/v2.12.0...v2.12.1) (2024-12-18)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [2.12.0](https://github.com/dcmjs-org/dcmjs/compare/v2.11.7...v2.12.0) (2024-12-16)
+
+### Features
+
+-   **adapters:** default to creation of labelmap images in 2D instead of a volume ([#1692](https://github.com/dcmjs-org/dcmjs/issues/1692)) ([960b75a](https://github.com/dcmjs-org/dcmjs/commit/960b75a2ebed3a141ceefcf4f38fb6de388a4ca0))
+
+## [2.11.7](https://github.com/dcmjs-org/dcmjs/compare/v2.11.6...v2.11.7) (2024-12-13)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.6](https://github.com/dcmjs-org/dcmjs/compare/v2.11.5...v2.11.6) (2024-12-12)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.5](https://github.com/dcmjs-org/dcmjs/compare/v2.11.4...v2.11.5) (2024-12-12)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.4](https://github.com/dcmjs-org/dcmjs/compare/v2.11.3...v2.11.4) (2024-12-12)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.3](https://github.com/dcmjs-org/dcmjs/compare/v2.11.2...v2.11.3) (2024-12-12)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.2](https://github.com/dcmjs-org/dcmjs/compare/v2.11.1...v2.11.2) (2024-12-11)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+## [2.11.1](https://github.com/dcmjs-org/dcmjs/compare/v2.11.0...v2.11.1) (2024-12-11)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [2.11.0](https://github.com/dcmjs-org/dcmjs/compare/v2.10.0...v2.11.0) (2024-12-11)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
+# [2.10.0](https://github.com/dcmjs-org/dcmjs/compare/v2.9.0...v2.10.0) (2024-12-11)
+
+**Note:** Version bump only for package @cornerstonejs/adapters
+
 # [2.9.0](https://github.com/dcmjs-org/dcmjs/compare/v2.8.6...v2.9.0) (2024-12-09)
 
 **Note:** Version bump only for package @cornerstonejs/adapters